% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathos.R
\name{plot_segments}
\alias{plot_segments}
\alias{plot_segments}
\alias{plot_arrows}
\title{Plot segments and arrows}
\usage{
plot_segments(d, ...)

plot_arrows(d, ...)
}
\arguments{
\item{d}{data frame}

\item{...}{arguments passed to [segments()]}
}
\description{
Add segments or arrows to plot from prepare_segments.
}
\details{
Input is a data frame with `x0`, `y0`, `x`, `y` columns as per [segments]
}
