% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathos.R
\name{prepare_path_segments}
\alias{prepare_path_segments}
\title{Paths as segments}
\usage{
prepare_path_segments(X, Y, G)
}
\arguments{
\item{X}{input X coordinate}

\item{Y}{input Y coordinate}

\item{G}{input grouping index}
}
\description{
Prepare paths of coordinates by Group as segments
}
\details{
This is really specific to X, Y, G inputs but can be made more general to any set of grouped (and implicitly arranged) records.
}
