% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathos.R
\name{prepare_segments}
\alias{prepare_segments}
\title{Segments}
\usage{
prepare_segments(x, scale = 3600, ...)
}
\arguments{
\item{x}{[raster::brick] with two layers of vector field components (u, v)}

\item{scale}{multiplication factor for u-v}

\item{...}{ignored}
}
\description{
Prepare segments from two-layer u-v brick.
}
\details{
The input is currently assumed to be a two-layer raster brick with horizontal (U) and vertical (V) components to vectors
at each pixel.
}
